requirements.txt
langchain-core>=0.1.0
httpx>=0.24.1
requests>=2.31.0
typing-extensions>=4.7.0
pydantic>=2.0.0
python-dotenv>=1.0.0
langchain>=0.3.0
langgraph>=0.0.26
